[FIX] Be consistent in how default values for fields should be assigned
In the fields section, we mention Default functions for fields should be
declared with a lambda call on self, instead of assigning a function.
However, in the models section, we provide an example where we assign a
function, which in addition forces us to move methods for default values
to be located above fields.