[IMP] add guidelines about website features
[Odoo 14.0 recommends running HttpCase tests on `post_install` mode only][1]. This can help on certain situations, but increments chances of some website modules polluting others.

One simple and common example are `website_sale` extensions in the OCA/e-commerce repo. If a module adds the phone as a required field and another module adds the VAT as required field, when both are installed and both test the checkout workflow, both will probably fail because none of them would be filling the other required field. Note that this is still possible without the `post_install` mode; this just increments the possibilites.

In any case, since Odoo v12, we have multiple websites. So, any module that alters Odoo upstream behavior should have those alterations disabled by default, expecting the user to enable them only in the website(s) they want.

This should make tests integrate properly out of the box, and should make all website addons more reliable, but it is not obvious, so I think this should be documented into a guideline.

[1]: https://github.com/odoo/odoo/blob/9b0578766709b27c91d85bf6246b4bc5e43ca27d/odoo/tests/common.py#L1377

@Tecnativa TT25963