2025 update to bylaws
Review of the bylaws done in between April 2025 and June 2025 by the Governance Working Group of the OCA.
Revision mainly aims to make it clearer and to incorporate some minor change that were not inline with the historical actual way the association has been operating.
As a reference the original bylaws page is at https://odoo-community.org/about/by-laws.

The most important changes are:

**Article 5.**

- Removed *"Any physical person may request to join: if (...) they are not salaried by the Association"*
condition as we now have people being paid buy the OCA, and we don't see a good reason for them no not be allowed to be Members.

- Removed *"Requests to become a member must be addressed to the Board. The Board admits new members and informs the Delegate Assembly."*
as this is not consistent to the current practice since the beginning on the association.

 **Article 11.**

- Added *"Decisions concerning the amendment of the Statutes and the dissolution of the Association must be approved by a two-third majority of the Delegate members present, **and at least half of the Delegates must be present and vote**"*
for a stronger protection to bylaws changes.

**Article 15**

- Removed *"An existing Board Member must have attended 80 % of the board meetings during its term to be a candidate for a re-election."* as preventing a board member to reapply seems unreasonable.
The Board can exclude a Board Member during its mandate if the Board Member misses 3 consecutive board meetings.
There can be justified personal reasons for a person to step down.
We should let Delegates make a decision with their votes if this person is worth being reelected.

**Article 22**

- Changed *"A signed CLA is required for every Delegate member and before an individual is given commit rights to an OCA project." rewording "commit rights" to "contribution " and better consider non code contributions.

- Removed *"CLAs may be submitted by traditional postal mail, fax, or by emailing a scan of the signed copy."* and let to the the board to determine the most appropriate way to collect CLAs.

**Article 26**

- Removed *"Should divergences exist between the translations of these articles, the French version shall prevail."* as we found that the original bylaws are already in English and no French version exists.